requirements.txt
aniso8601==8.0.0
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-Dropzone==1.5.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==7.0.0
pymongo==3.10.1
pytesseract==0.3.1
pytz==2019.3
six==1.14.0
visitor==0.1.3
Werkzeug==0.16.0
